fix: 검색페이지 버그 수정(DB 연결실패 이슈)
